experimental: relocate circuits/surgery to qldpc.experimental.surgery
Move the lattice-surgery module out of circuits/ into a new top-level experimental subpackage, marking its API unstable. git mv preserves history; the four co-located tests move with it and their absolute qldpc.circuits.surgery imports (and the notebook's) are repointed. Add a docstring-only experimental/__init__.py namespace that does not import surgery, so 'import qldpc' still does not pull in surgery/stim. No packaging/mypy/coverage/docs config needed changing (all auto-discover under src/qldpc). All 183 surgery tests pass.
