+++ src/qldpc/experimental/__init__.py
@@ -1,0 +1,7 @@
+"""Experimental ``qldpc`` subpackages.
+
+EXPERIMENTAL: everything under ``qldpc.experimental`` is under active
+development. Public APIs here are unstable and may change without notice or
+deprecation. Importing ``qldpc`` does not import these subpackages; import them
+explicitly, e.g. ``import qldpc.experimental.surgery``.
+"""
