+++ src/qldpc/experimental/surgery/gadget.py
